LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/stale-while-revalidate/stale-css-expected.txt


FAIL Cache returns stale resource assert_equals: (revalidated version loaded) expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
